split sale_order.py in smaller bits, some flake8 cleanup
