+++ sale_quotation_sourcing/models/__init__.py
@@ -19,3 +19,4 @@
 #
 
 from . import sale_order
+from . import procurement
