+++ sale_quotation_sourcing/wizard/__init__.py
@@ -17,5 +17,4 @@
 #    along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #
 #
-import models
-import wizard
+from . import quotation_sourcing
